Fixing the SVG transform-origin during initial mount (#2949)
Dimension is measured on the client side. Therefore, the origin cannot be measured on the first mount, resulting in a jump. Before the dimension is measured, set "transformBox" to "Fill Box" and center the origin to leave the initial origin to the browser.

This behavior is only valid when manually calculating the last value of "transformOrigin".